fix minimization of symbolic dfa.
There is an issue when the automaton does not have any
accepting states, but several transitions among states.
The result is that, after the greatest-fixpoint computation,
during the building of the automaton, the algorithm tried
to add a non-deterministic transition, which make the
procedure to fail.

Fix that by normalizing the transitions before building the
automaton. "Normalize" here means to have only one guard
per edge.